Error Plugin v0.6 Adds displaying errors in admin area.
Also includes major change to DB structure.
Serialized arrays are not saved as base64 encoding,
and their fields have been changed to MEDIUMTEXT.
This is a beaking change, but neccessary for fixing
encoding issues with the raw serialization.

TODO: Improved version of JS needs to be incorporated into debug template.